Initial Project

+++ frontend_admin/src/pages/admin/category/index.jsx
@@ -12,6 +12,7 @@ const CategoryPage = () => {
   const itemsPerPage = 10;
   const [currentPage, setCurrentPage] = useState(1);
   const [filteredCategories, setFilteredCategories] = useState([]);
+  const [searchTerm, setSearchTerm] = useState('');
 
   useEffect(() => {
     dispatch(fetchAllCategories());
@@ -24,14 +25,38 @@ const CategoryPage = () => {
   }, [currentPage, categories]);
 
   const handleDelete = (id) => {
-    dispatch(deleteCategoryById(id))
-      .then(() => {
-        SweetAlert.success('Success', 'Item Remove Category');
-      })
-      .catch(() => {
-        SweetAlert.error('Error', 'Failed to remove item from Category');
-      });
+    if (window.confirm('Are you sure you want to delete this record?')) {
+      dispatch(deleteCategoryById(id))
+        .then(() => {
+          SweetAlert.success('Success', 'Item Remove Category').then(() =>
+            dispatch(fetchAllCategories())
+          );
+        })
+
+        .catch(() => {
+          SweetAlert.error('Error', 'Failed to remove item from Category');
+        });
+    }
   };
+
+  const handleSearch = (e) => {
+    setCurrentPage(1);
+    setSearchTerm(e.target.value);
+  };
+
+  const filteredData = categories.filter((category) => {
+    return (
+      category.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
+      category.description.toLowerCase().includes(searchTerm.toLowerCase())
+    );
+  });
+
+  const pageCount = Math.ceil(filteredData.length / itemsPerPage);
+  const displayedCategories = filteredData.slice(
+    (currentPage - 1) * itemsPerPage,
+    currentPage * itemsPerPage
+  );
+  const currentBlock = Math.ceil(currentPage / 5);
 
   if (loading) {
     return <h1>Loading</h1>;
@@ -42,19 +67,19 @@ const CategoryPage = () => {
       <div className="page-title">
         <div className="row">
           <div className="col-12 col-md-6 order-md-1 order-last">
-            <h3>category</h3>
+            <h3>Category</h3>
             {error && (
               <div
                 className="alert alert-danger alert-dismissible fade show"
                 role="alert"
               >
-                {error}
+                {error.message}
                 <button
                   type="button"
                   className="btn-close"
                   data-bs-dismiss="alert"
                   aria-label="Close"
-                ></button>
+                />
               </div>
             )}
           </div>
@@ -67,9 +92,7 @@ const CategoryPage = () => {
                 <li className="breadcrumb-item">
                   <a href="index.html">Dashboard</a>
                 </li>
-                <li className="breadcrumb-item active" aria-current="page">
-                  Category
-                </li>
+                <li className="breadcrumb-item active" aria-current="page"></li>
               </ol>
             </nav>
           </div>
@@ -87,21 +110,33 @@ const CategoryPage = () => {
               <i className="fas fa-user"></i> Add Category
             </button>
           </div>
+
           <div className="card-body">
+            <div className="mb-3">
+              <input
+                type="text"
+                className="form-control"
+                placeholder="Search by name or description"
+                value={searchTerm}
+                onChange={handleSearch}
+              />
+            </div>
             <table className="table table-striped" id="table1">
               <thead>
                 <tr>
                   <th>No</th>
                   <th>Nama Kategori</th>
+                  <th>Description</th>
                   <th>Image</th>
                   <th>Status</th>
                 </tr>
               </thead>
               <tbody>
-                {filteredCategories.map((row, index) => (
+                {displayedCategories.map((row, index) => (
                   <tr key={row.id}>
                     <td>{(currentPage - 1) * itemsPerPage + index + 1}</td>
                     <td>{row.name}</td>
+                    <td>{row.description}</td>
                     <td>
                       <img
                         src={row.image_category}
@@ -128,7 +163,6 @@ const CategoryPage = () => {
                 ))}
               </tbody>
             </table>
-
             <nav aria-label="Page navigation example">
               <ul className="pagination pagination-primary">
                 <li
@@ -141,12 +175,25 @@ const CategoryPage = () => {
                     Previous
                   </span>
                 </li>
-                <li className="page-item">
-                  <span className="page-link active">Page {currentPage}</span>
-                </li>
+                {[...Array(pageCount > 5 ? 5 : pageCount)].map((_, index) => (
+                  <li key={index} className="page-item">
+                    <span
+                      className={`page-link ${
+                        currentPage === index + 1 + (currentBlock - 1) * 5
+                          ? 'active'
+                          : ''
+                      }`}
+                      onClick={() =>
+                        setCurrentPage(index + 1 + (currentBlock - 1) * 5)
+                      }
+                    >
+                      {index + 1 + (currentBlock - 1) * 5}
+                    </span>
+                  </li>
+                ))}
                 <li
                   className={`page-item ${
-                    currentPage * itemsPerPage >= categories.length
+                    currentPage * itemsPerPage >= filteredData.length
                       ? 'disabled'
                       : ''
                   }`}
